Refactor travel model, move gas station logic to map
Renamed TravelQueryModel properties (Origin/Destination to Orig/Dest) and switched coordinates to double. Removed GasStationsViewer and integrated gas station display into MapComponent with new route and point loading logic. Updated GeoJSON models, OSRM router, and config files. Cleaned up unused code.

+++ src/Libs/GasStationPrices/ViewModels/Place.cs
@@ -3,6 +3,6 @@
 public record class Place
 {
     public required string Address { get; set; }
-    public required float Latitude { get; set; }
-    public required float Longitude { get; set; }
+    public required double Latitude { get; set; }
+    public required double Longitude { get; set; }
 }
